build: add libopus via FetchContent + fix uint32_t DBG ambiguity
- FetchContent libopus v1.5.2 (static, no tests)
- Added opus include dir and link target to SharedCode/Mix2Go
- Cast seq to (int) in DBG to resolve JUCE String << ambiguity
- VST3/AU/Standalone confirmed built successfully with opus_encode linked

+++ source/Streaming/AudioStreamManager.h
@@ -276,7 +276,7 @@ private:
 
         if ((seq % 100) == 0)
         {
-            DBG("[Mix2Go] Stats: seq=" << seq
+            DBG("[Mix2Go] Stats: seq=" << (int)seq
                 << "  sent=" << (int)m_sender.getPacketsSent()
                 << "  fifoLevel=" << m_fifo.getNumReady()
                 << "  overruns=" << (int)m_fifo.getOverrunCount()
